==> Part 1 - Data Cleaning - Movie Titles.ipynb <==
import pandas as pd
import numpy as np
import re

### Trying to read the 'movie_titles.csv' gives an error.

file = '/raid/datasets/misc/netflix-prize-data/movie_titles.csv'

movie_titles = pd.read_csv('movie_titles.csv')

### Looking at line 72, we see that there are commas in titles which confuses the csv reader. Hence the error

!sed '72q;d' 'movie_titles.csv'  

### The idea now is to find a pattern in commas that we can exploit and change the seperator from comma to something else (I chose semicolon). First we read the file using standard python open() method and check if the number of lines we get are the same as in the file

with open(file,'r',encoding='utf-8',errors='ignore') as f:
    data = f.readlines()

!wc -l movie_titles.csv

len(data)

### Reading like above adds the newline character to every line. Removing it using strip() method of strings, we get the data as a list. (strip() drops any white space from beginning or end of a string)

data = [d.strip() for d in data]

data[:5]

### First we check that our chosen seperator is not in the dataset. If it is, we cannot use it. 

for d in data:
    f=re.search(';',d)
    if f is not None:
        print(f)

### Looking at the five lines, we can see that the first two commas in every line are good candidates for being a seperator, so we use python's regex to replace first 2 commas to semicolon like so:

new_data=[]
for d in data:
    new_data.append(re.sub(',',';', d,2))

new_data[70:75]

with open('movie_titles_tmp.csv','w') as f:
    for d in new_data:
        f.write(d+'\n')

### Now we can try to read the dataset again, using semicolon as seperator

df=pd.read_csv('movie_titles_tmp.csv',sep=';',header=None,names=['movie_id','year','title'],index_col='movie_id')

df.head()

### We want to convert the year from float to int. But we get an error

df.year=df.year.astype('int32')

### Apparently there is an NA or Inf and python cannot convert to integer. So we check the unique values of year column

df.year.unique()

### We see that there is a 'nan'. Then we try to filter out in the dataset and find a handful columns.

df[df.year.isna()]

### For now we just drop the columns. First we find the indices of these columns and then use pandas drop() method to delete them

year_na = df[df.year.isna()].index
df.drop(year_na,axis=0,inplace=True)

### Finally we check again that there is no nan in the year column

df[df.year.isna()]

df.year=df.year.astype('int32')

df.head()

### Finally we save the clean dataset

df.to_csv('movie_titles_clean.csv')
