==> Part 2 - Data Cleaning - Ratings.ipynb <==
import pandas as pd
import numpy as np
import re, os
from pathlib import Path
from tqdm import tqdm
import pyarrow as pa
import pyarrow.parquet as pq

### Looking at the first few lines we notice that a peculiar structure. There are two types of lines. One with just the movie_id followed by a colon. And other with user_id, rating and date for that movie_id.

file1 = Path('/raid/datasets/misc/netflix-prize-data/combined_data_1.txt')
file2 = Path('/raid/datasets/misc/netflix-prize-data/combined_data_2.txt')
file3 = Path('/raid/datasets/misc/netflix-prize-data/combined_data_3.txt')
file4 = Path('/raid/datasets/misc/netflix-prize-data/combined_data_4.txt')

with open(file1,'r',encoding='utf-8',errors='ignore') as f:
    data = f.readlines()

data[:10]

### Loading a text file, followed by stripping the newline character

data = [d.strip() for d in data]

data[:11]

### First we make a list of all movie_ids and compare them with the movie_ids from first part.

movie_list=[]
for i,d in tqdm(enumerate(data),total=len(data)):
    f=re.search('\d+:',d)
    if f is not None:
#         print(f'{i} : {d}')
        movie_list.append(d[:-1])

movie_list[:5]

### Loading movie_ids from first part

movie_data = pd.read_csv('movie_titles_clean.csv',index_col='movie_id')

movie_data.head()

movie_data.index

movie_list_int = [int(d) for d in movie_list]

### Here we take the difference of both list of movie_ids. We notice that while there are many movies that don't have a rating (first command); there is only one movie in the rating list that doesn't have movie information (second command)

set(movie_data.index).difference(movie_list_int)

set(movie_list_int).difference(movie_data.index)

### In the following we read the dataset. First we search for the movie_id line. When we find such a line (f!=None), we store the id in 'movie' variable. When we don't have such a line, its the rating line. We split it on comma and combine it with movie_id.

rating_data=[]
for i,d in tqdm(enumerate(data),total=len(data)):
    f=re.search('\d+:',d)
    if f is not None:
        movie = d[:-1]
        continue
    rating=d.split(',')
    rating_data.append((movie,*rating))

### Now we have list of tuples with the format (movie_id, user_id, rating, date). We make a pandas dataframe from it. 

rating_df = pd.DataFrame(rating_data,columns=['movie_id','user_id','rating','date'])

rating_df.head()

rating_df.user_id.nunique()

rating_df.movie_id.nunique()

### Dropping the movie for which info is not available

rating_df = rating_df.query('movie_id!=4388')

### Here we ensure that there is only ony rating for each movie_id and user_id by first grouping the data by movie_id, user_id and counting the number of ratings.

uniq_rating_df=rating_df.groupby(['movie_id','user_id'])['rating'].count()

### Checking if we have only 1 rating for each combination

uniq_rating_df[uniq_rating_df!=1]

rating_df.index.name = 'index'

# rating_df.to_csv('rating.csv')
rating_df = rating_df.drop('date',axis=1).apply(pd.to_numeric)

### Doing the above for other three files

with open(file2,'r',encoding='utf-8',errors='ignore') as f:
    data2 = f.readlines()
print(len(data2))
with open(file3,'r',encoding='utf-8',errors='ignore') as f:
    data2.extend(f.readlines())
print(len(data2))
with open(file4,'r',encoding='utf-8',errors='ignore') as f:
    data2.extend(f.readlines())
print(len(data2))

data2=[d.strip() for d in data2]

rating_data2=[]
for i,d in tqdm(enumerate(data2),total=len(data2)):
    f=re.search('\d+:',d)
    if f is not None:
        movie = d[:-1]
        continue
    rating=d.split(',')
    rating_data2.append((movie,*rating))

rating2_df = pd.DataFrame(rating_data2,columns=['movie_id','user_id','rating','date'])

rating2_df = rating2_df.drop('date',axis=1).apply(pd.to_numeric)

### Concatenating Datasets

all_rating_df=pd.concat([rating_df,rating2_df])

all_rating_df.index.name = 'index'

all_rating_df.head()

table = pa.Table.from_pandas(all_rating_df)

pq.write_table(table, 'all_ratings.parquet')

all_rating_df.to_csv('all_rating.csv')
