==> Part 3 - PySpark-Recommendation.ipynb <==
from pyspark.ml.recommendation import ALS
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.sql import SparkSession
from pyspark.sql.types import IntegerType

### Adjust spark.driver.memory for your system. You can leave it out, but if you get Java Out of Memory error, you need to put it in there.

spark = SparkSession.builder.appName('Recommendation_system').config('spark.executor.memory', '5g').config('spark.driver.memory','120g').config('spark.local.dir','/home/haris/raid/tmp').getOrCreate()
# spark = SparkSession.builder.appName('Recommendation_system').getOrCreate()

### Here you get a UI to monitor spark cluster status

spark

df = spark.read.csv('all_rating.csv',header=True)

df.show(10)

### ALS expects integer values for user_id, movie_id and rating

als_data=df.select(df['movie_id'],df['user_id'],df['rating'])
als_data = als_data.withColumn("movie_id", als_data["movie_id"].cast(IntegerType()))
als_data = als_data.withColumn("user_id", als_data["user_id"].cast(IntegerType()))
als_data = als_data.withColumn("rating", als_data["rating"].cast(IntegerType()))

(training,test)=als_data.randomSplit([0.8, 0.2])

als = ALS(userCol="user_id",itemCol="movie_id",ratingCol="rating",coldStartStrategy="drop",nonnegative=True,implicitPrefs=False)
# als = als.setMaxIter(5).setRegParam(0.09).setRank(25)

evaluator=RegressionEvaluator(metricName="rmse",labelCol="rating",predictionCol="prediction")

### Use following three lines if you want to run a single model fitting. Otherwise continue for Parameter Grid Search

model=als.fit(training)

predictions=model.transform(test)
rmse=evaluator.evaluate(predictions)

print("RMSE="+str(rmse))

### Parameter Grid Search 

from pyspark.ml.tuning import ParamGridBuilder
from pyspark.ml.tuning import CrossValidator

param_grid = ParamGridBuilder().addGrid(als.regParam, [.05, .1, 1.5]).build() # .addGrid(als.maxIter, [5, 250, 500]).addGrid(als.rank, [5, 80, 120])

cv = CrossValidator(estimator = als, estimatorParamMaps = param_grid, evaluator = evaluator, numFolds = 5)

model = cv.fit(training)

best_model = model.bestModel

predictions=best_model.transform(test)
rmse=evaluator.evaluate(predictions)

print ("**Best Model**")
print (f"RMSE = {rmse}")
print (f" Rank: {best_model.rank}")
print (f" MaxIter: {best_model._java_obj.parent().getMaxIter()}")
print (f" RegParam: {best_model._java_obj.parent().getRegParam()}")
