==> Part 4 - sklearn-Recommendation.ipynb <==
import numpy as np
from sklearn.metrics import pairwise_distances, pairwise_distances_chunked
from scipy.spatial.distance import cosine
import pandas as pd
from scipy import sparse
from sklearn.metrics.pairwise import cosine_similarity
import sklearn
import pyarrow as pa
import pyarrow.parquet as pq
from tqdm import tqdm

df = pq.read_table('all_ratings.parquet').to_pandas()

df.head()

### Add indices to users and movies

num_movies = len(df.movie_id.unique())
num_users = len(df.user_id.unique())

num_movies, num_users

movie_to_idx = {m:idx for idx,m in enumerate(df.movie_id.unique())}
idx_to_movie = {idx:m for idx,m in enumerate(df.movie_id.unique())}

user_to_idx = {m:idx for idx,m in enumerate(df.user_id.unique())}
idx_to_user = {idx:m for idx,m in enumerate(df.user_id.unique())}

df['movie_idx']=df.movie_id.map(movie_to_idx)
df['user_idx']=df.user_id.map(user_to_idx)

df.shape

df.head()

table = pa.Table.from_pandas(df)
pq.write_table(table, 'all_ratings_with_indices.parquet')

## Loading ratings with indices and converting to sparse format

df = pq.read_table('all_ratings_with_indices.parquet',columns=['rating','user_idx','movie_idx']).to_pandas()

df.shape

df.head()

num_users = df.user_idx.max()+1
num_movies = df.movie_idx.max()+1

num_movies, num_users

All = sparse.csr_matrix((df.rating,(df.user_idx,df.movie_idx)),shape=(num_users,num_movies))
All

### Selecting only 30,000 users out of 480,189

A = All[:30000,:]

# user_sim = pairwise_distances(A, metric="cosine",n_jobs=-1)

### I ran out of memory with 256GB RAM! So I will now use **pairwise_distances_chunked** which is meant for limited memory

user_sim = pairwise_distances_chunked(A, metric="cosine",n_jobs=-1)

def get_neighbors(user_sum,k=5):
    neighbors = []
    for chunk in tqdm(user_sim):
        idx = np.argpartition(chunk, range(k),axis=1)[:,1:k]
        neighbors.append(idx)
    return np.concatenate(neighbors)

recc = get_neighbors(user_sim)

### The following table shows the user_ids of 4 closest users to all 30,000 users we selected above

recc

recc.shape
